% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fit.R
\name{add_FIT}
\alias{add_FIT}
\title{Function to add \%FIT to stats}
\usage{
add_FIT(file = file.choose(), verbose = F)
}
\arguments{
\item{file}{model data.frame}

\item{verbose}{display additional information}
}
\description{
Function to read, add \%FIT and write stats
}
