% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_var}
\alias{calculate_var}
\alias{calculate_DZ}
\alias{calculate_CWP}
\alias{calculate_CDNC}
\alias{calculate_column}
\alias{calculate_ACC_RAIN}
\title{Functions to calculate specific variables to model evaluation}
\usage{
calculate_DZ(file, file2 = file, as_raster = F, verbose = T)

calculate_CWP(file, met, verbose = T, ...)

calculate_CDNC(file, met, verbose = T, ...)

calculate_column(file, var, met, verbose = F, ...)

calculate_ACC_RAIN(file, map = file, last = TRUE, verbose = T)
}
\arguments{
\item{file}{file name}

\item{file2}{file name (for PH)}

\item{verbose}{set TRUE to display additional information}

\item{met}{file with meteorological variables (ALT, PH,PHBD)}

\item{...}{passed to calculate_DZ}

\item{var}{variable name}

\item{map}{a file with LAT, LON and other grid information}

\item{last}{calculate using only the last time}
}
\value{
a raster object
}
\description{
function read and calculate model layer thickness

functions read and to calculate Cloud water path

functions read and to calculate Cloud droplet number concentration

functions read and to calculate column sum concentration

functions read and to calculate column sum concentration
}
