% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Filter table A to perform a fair comparison with table B}
\usage{
comparison(tabA, tabB, use_names, verbose = T)
}
\arguments{
\item{tabA}{Table A, from evaluate}

\item{tabB}{Table B, from evaluate}

\item{verbose}{display additional information}

\item{Use_names}{list of names to be used instead of Table B}
}
\value{
Table A containing only the row.names in table B
}
\description{
Filter table A to perform a fair comparison with table B using row.names
}
\note{
to perform a comparison check both tables using comparison(A,B) and comparison(B,A)
}
\examples{
model <- 1:10
data  <- model + rnorm(10,0.2)
A     <- stats(mo = model, ob = data)
for(i in 2:10){
   data  <- model + rnorm(10,0.2)
   A     <- rbind(A,stats(mo = model, ob = data))
}
row.names(A) <- c('a0','a1','a3','b0','b1','b2','b3','d0','d1','d2')

data   <- model + rnorm(10,0.2)
B      <- stats(mo = model, ob = data)
for(i in 2:5){
    data   <- model + rnorm(10,0.2)
    B      <- rbind(B,stats(mo = model, ob = data))

}
row.names(B) <- c('a0','a1','z0','z1','z2')

comparison(A,B)
comparison(B,A)
}
