% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacement.R
\name{displace}
\alias{displace}
\title{Displace a vector}
\usage{
displace(x, n = 0)
}
\arguments{
\item{x}{vector}

\item{n}{number of displacement (positive is right)}
}
\description{
Displace a vector
}
