% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Model statistical evaluation}
\usage{
evaluate(mo, ob, station, table = NULL, clean = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{mo}{model data.frame}

\item{ob}{observed data.frame}

\item{station}{name of the station}

\item{table}{a data.frame with output from evaluate or stats}

\item{clean}{remove rows with zero observations}

\item{verbose}{display additional information}

\item{...}{arguments to be passing to stats}
}
\description{
Statistical evaluation
}
\examples{
model <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/model.Rds"))
obs   <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/obs.Rds"))

# first a test with no observed data
table <- evaluate(mo = model, ob = obs, station = "VVIbes")
# now a test with a few observed values
table <- evaluate(mo = model, ob = obs, station = "Americana", table = table)
# new tests with no data will be discated
table <- evaluate(mo = model, ob = obs, station = "VVIbes", table = table)
# if the first evaluation has no data, the last call can remove the line
table <- evaluate(mo = model, ob = obs, station = "Americana", table = table, clean = TRUE)

}
