% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_8h_max.R
\name{extract_max_8h}
\alias{extract_max_8h}
\title{Creata a NetCDF file with the surface maximum of O3}
\usage{
extract_max_8h(
  filelist,
  variable = "o3",
  field = "4d",
  prefix = "max_8h",
  units = "ug m-3",
  meta = T,
  verbose = TRUE
)
}
\arguments{
\item{filelist}{list of files to be read}

\item{variable}{variable name}

\item{field}{'4d' (default), '3d', '2d' or '2dz' see notes}

\item{prefix}{to output file, defolt is serie}

\item{units}{units on netcdf file (default is ug m-3), change to skip unit conversion}

\item{meta}{use Times, XLONG and XLAT data (only work with 2d variable for file)}

\item{verbose}{display additional information}
}
\description{
Read the values from o3 and T2, convert o3 to ug m-3 and calculate the maximum of 8-hour moving avarage from a list of files.
}
\note{
The field argument '4d' / '2dz' is used to read a 4d/3d variable droping the 3rd dimention (z).
}
\examples{
dir.create(file.path(tempdir(), "MAX"))
folder <- system.file("extdata",package="hackWRF")
wrf_file <- paste0(folder,"/wrf.day1.o3.nc")
extract_max(filelist = wrf_file,prefix = paste0(file.path(tempdir(),"MAX"),'/mean'))

}
