% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_regime.R
\name{extract_regime}
\alias{extract_regime}
\title{Creata a NetCDF file with the ratio of FORM/(NO2 + O3)}
\usage{
extract_regime(
  filelist,
  variable_mon = "FORM",
  variable_dem = c("no2", "o3"),
  field = "4d",
  prefix = "ratio",
  units = "",
  meta = T,
  filename,
  verbose = TRUE
)
}
\arguments{
\item{filelist}{list of files to be read}

\item{variable_mon}{variable name for nominador}

\item{variable_dem}{variable name for denominador}

\item{field}{'4d' (default), '3d', '2d' or '2dz' see notes}

\item{prefix}{to output file, defolt is serie}

\item{units}{units on netcdf file (default is ppmv)}

\item{meta}{use Times, XLONG and XLAT data (only work with 2d variable for file)}

\item{filename}{name for the file, in this case prefix is not used}

\item{verbose}{display additional information}
}
\description{
Read and calculate the mean value of a variable from a list of wrf output files.
}
\note{
The field argument '4d' / '2dz' is used to read a 4d/3d variable droping the 3rd dimention (z).

ratio < 0.28, VOC-limited, +VOC emissions -> +O3 concentration

ratio > 0.28, NOx-limited, +NOx emissions -> +O3 concentration
}
\references{
Sillman, S. (1995). The use of NO y, H2O2, and HNO3 as indicators
for ozone‐NO x‐hydrocarbon sensitivity in urban locations. Journal of
Geophysical Research: Atmospheres, 100(D7), 14175-14188.
}
