% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_serie.R
\name{extract_serie}
\alias{extract_serie}
\title{Extract time series of wrf file list of lat/lon}
\usage{
extract_serie(
  filelist,
  point,
  variable = "o3",
  field = "4d",
  prefix = "serie",
  new = TRUE,
  return.nearest = F,
  fast = F,
  verbose = TRUE
)
}
\arguments{
\item{filelist}{list of files to be read}

\item{point}{data.frame with lat/lon}

\item{variable}{variable name}

\item{field}{'4d' (defoult), '3d', '2d' or '2dz' see notes}

\item{prefix}{to output file, defolt is serie}

\item{new}{start a new file (defoult)}

\item{return.nearest}{return tha data.frame of nearest points instead of extract the serie}

\item{fast}{faster calculation of grid distances but less precise}

\item{verbose}{display additional information}
}
\description{
Read and extract data from a list of wrf output files and a table of lat/lon points based on the distance of the points and the center of model grid points, points outside the domain (and points on domain boundary) are not extracteds.
}
\note{
The field argument '4d' or '2dz' is used to read a 4d/3d variable droping the 3rd dimention (z).
}
\examples{
cat('Example 1: INMET stations for 2015\n')
stations <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/inmet_2015.Rds"))

cat('Example 2: METAR stations of Brazil\n')
stations <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/metar-br.Rds"))

cat('Example 3: Brazilian Air QUality: CETESB (SP), RAMQAr (ES) and SMAC (RJ)\n')
stations <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/stations.Rds"))

files    <- dir(path = system.file("extdata",package="hackWRF"),pattern = 'wrf',full.names = TRUE)
dir.create(file.path(tempdir(),"SERIE"))
folder <- file.path(tempdir(),"SERIE")
# extract data for 3 locations
extract_serie(filelist = files, point = stations[1:3,],prefix = paste0(folder,'/serie'))

}
