% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_serie_ij.R
\name{extract_serie_ij}
\alias{extract_serie_ij}
\title{Extract time series of wrf file list of ij}
\usage{
extract_serie_ij(
  filelist,
  point,
  variable = "o3",
  field = "4d",
  prefix = "serie",
  new = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{filelist}{list of files to be read}

\item{point}{data.frame with i and j position}

\item{variable}{variable name}

\item{field}{'4d' (defoult), '3d', '2d' or '2dz' see notes}

\item{prefix}{to output file, defolt is serie}

\item{new}{start a new file (defoult)}

\item{verbose}{display additional information}
}
\description{
Read and extract data from a list of wrf output files and a table of i and j points of model grid.
}
\note{
The field argument '4d' or '2dz' is used to read a 4d/3d variable droping the 3rd dimention (z).
}
\examples{
stations <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/stations.Rds"))
files    <- dir(path = system.file("extdata",package="hackWRF"),pattern = 'wrf',full.names = TRUE)
dir.create(file.path(tempdir(),"SERIE"))
folder <- file.path(tempdir(),"SERIE")
extract_serie(filelist = files, point = stations,prefix = paste0(folder,'/serie'))

}
