% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{first}
\alias{first}
\title{Get the first element of a vector of data.frame}
\usage{
first(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
first element
}
\description{
This function return the first member of the input
}
\examples{
a <- 666:1
print(first(a))

}
