% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\name{format_stats}
\alias{format_stats}
\title{Function to format stat files}
\usage{
format_stats(
  file = file.choose(),
  select,
  rename = NULL,
  digits = NULL,
  multiplier = NULL,
  suffix = "_formatted",
  verbose = F
)
}
\arguments{
\item{file}{data.frame with stats}

\item{select}{name of the column}

\item{rename}{vector with new name of each column}

\item{digits}{vector of digits for each column, "i" for integer}

\item{multiplier}{multiplier to change units}

\item{suffix}{file suffix}

\item{verbose}{display additional information}
}
\description{
Function to read, add \%FIT and write stats
}
\note{
rename can contains NA, for each NA the name don't change
}
