% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Custom function to interpolation of a raster from satellite to a raster from model}
\usage{
interpolate(x, y, fast = F, e = extent(-80, -25, -40, 10), verbose = T)
}
\arguments{
\item{x}{a raster from observation}

\item{y}{a raster from model}

\item{fast}{to use only the raster extent region}

\item{e}{raster extenet}

\item{verbose}{set TRUE to display additional information}
}
\value{
a raster object
}
\description{
function to interpolate an satellite observation to model domain
}
\seealso{
for raster from satellite observations see \code{\link{read_satellite}}. For raster from model see wrf_raster from eixport package and \code{\link{calculate_var}}.
}
