% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Get the last element of a vector}
\usage{
last(x)
}
\arguments{
\item{x}{vector}
}
\value{
last element
}
\description{
This function return the last member of a vector
}
\examples{
a <- 1:666
print(last(a))

}
