% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon.R
\name{latlon}
\alias{latlon}
\title{Get latitude and longitude.}
\usage{
latlon(data, coord = "BR-AQ", proj = "+proj=longlat", verbose = T)
}
\arguments{
\item{data}{data.frame with row.names of correscponding locations}

\item{coord}{data.frame or 'BR-AQ','BR-METAR','BR-INMET' containing lat, lon and row.names of each location}

\item{proj}{projection string to CRS / proj4string (sp package), defoult in "+proj=longlat"}

\item{verbose}{to display additional information}
}
\description{
Get latitude and longitude.
}
\examples{
sample <- read.stat(paste0(system.file("extdata", package = "hackWRF"),
                           "/sample.csv"),verbose=TRUE)
row.names(sample) <- c("Americana","Campinas","Congonhas")

stations <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/stations.Rds"))

coord    <- latlon(sample, coord = stations)
print(coord)

}
