% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDA8.R
\name{MDA8}
\alias{MDA8}
\title{Maximum Daily 8-hr Average}
\usage{
MDA8(x, maximum = F)
}
\arguments{
\item{x}{data.frame with time column and aditional columns to be processed}

\item{maximum}{true to return the dayli maximum instead of the hourly 8-hr}
}
\value{
data.frame
}
\description{
function to calculate Ozone Maximum Daily 8-hr Average or 8-hr moving Average for a data.frame
}
