% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\title{Read and write metadata on a NetCDF file}
\usage{
meta(file = NA, var = 0, att = NA, action = "read", value = NA, verbose = F)
}
\arguments{
\item{file}{file name}

\item{var}{variable name, 0 to global and "?" to ask all names}

\item{att}{attribute names (NA for get all attnames)}

\item{action}{Read or write attribute, get return the value}

\item{value}{value to write}

\item{verbose}{display additional information}
}
\description{
Read and write metadata information of a NetCDF files
}
\examples{
nc <- paste0(system.file("extdata",package="hackWRF"),"/small.nc")
meta(nc)
meta(nc,att = "Title")
meta(nc,var = "?")
meta(nc,var = "so2")
meta(nc,var = "so2",att = "long_name")

}
