% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdump.R
\name{ncdump}
\alias{ncdump}
\title{Print a 'ncdum -h' command}
\usage{
ncdump(file = file.choose())
}
\arguments{
\item{file}{file name}
}
\description{
Read a NetCDF and print the medatada
}
\examples{
ncdump(file = paste0(system.file("extdata",package="hackWRF"),'/wrfinput_d01'))

}
