% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{nd}
\alias{nd}
\title{return character with zero left}
\usage{
nd(x, dig = 2)
}
\arguments{
\item{x}{numerical input}

\item{dig}{number of digits (defoult is 2)}
}
\value{
character with zeros left
}
\description{
return character with zero left
}
\examples{
a <- 1:66
print(nd(a))

}
