% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plot_diff}
\title{Plot raster difference}
\usage{
plot_diff(
  a,
  b,
  file_a,
  file_b,
  var = "",
  map,
  operation = mean,
  col,
  col_1,
  col_2,
  title = "",
  units_1 = expression(kton ~ yr^-1),
  units_2 = expression("\%"),
  lim_1 = NA,
  lim_2 = NA,
  lines_1,
  lines_2,
  save,
  w = 800,
  h = 1200,
  pt = 18,
  int = 10,
  grid_int = 15,
  grid_lat_min = -80,
  grid_lat_max = 80,
  force_max,
  ndig = 2,
  ...
)
}
\arguments{
\item{a}{raster A}

\item{b}{raster B}

\item{file_a}{file for raster A}

\item{file_b}{file for raster B}

\item{var}{sub title 1 and variable to be open for file_a and file_b}

\item{map}{map passed to eixport::wrf_raster}

\item{operation}{function passed to calc, default is mean}

\item{col}{color scale}

\item{col_1}{color for primary map}

\item{col_2}{color for secondary map and grid}

\item{title}{main title}

\item{units_1}{units suptitle for absolute difference}

\item{units_2}{units suptitle for relative difference}

\item{lim_1}{limits for absolute difference}

\item{lim_2}{limits for relative difference}

\item{lines_1}{primary map (sf class lines)}

\item{lines_2}{secondary map (sf class lines)}

\item{save}{TRUE to save with automatic name, or a filename}

\item{w}{width for png function}

\item{h}{height for png function}

\item{pt}{fond size for png function}

\item{int}{interval for latitude and logitude axis}

\item{grid_int}{int passed to project_grid function}

\item{force_max}{module of upper and lower limits for percentage plot}

\item{ndig}{number of digits for lengend_range}

\item{...}{arguments passed to raster_plot}

\item{grid_min_lat}{min_lat passed to project_grid function}

\item{grid_max_lat}{max_lat passed to project_grid function}
}
\description{
Plot difference (absolute and relative) of A - B single layer raster objects
}
\examples{
us    <- sf::read_sf(paste0(system.file("extdata",package="hackWRF"),"/us.shp"))
coast <- sf::read_sf(paste0(system.file("extdata",package="hackWRF"),"/coast.shp"))
A     <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/rasterA.Rds"))
B     <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/rasterB.Rds"))

plot_diff(a = A, b = B,
          lines_1 = coast, lines_2 = us,
          var = 'E_PM25J', title = 'A - B',
          lim_1 = 0.00001,lim_2 = 1)

}
