% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_raster}
\alias{plot_raster}
\title{Plot raster object}
\usage{
plot_raster(
  r,
  log = FALSE,
  min = -3,
  max,
  legend.shrink = 0.98,
  legend.width = 3,
  axe = !llaxis,
  llaxis = TRUE,
  int = 10,
  proj = FALSE,
  ...
)
}
\arguments{
\item{r}{raster}

\item{log}{TRUE to plot in log-scale}

\item{min}{minimum for log plot (default is -3)}

\item{max}{maximum for log plot}

\item{legend.shrink}{legend height (default is 0.98)}

\item{legend.width}{legend width (default is 3)}

\item{axe}{to plot axis}

\item{llaxis}{to plot custom axis}

\item{int}{argument passed to latitude / longitude functions}

\item{proj}{TRUE to project the raster to latlon}

\item{...}{arguments to be passing to stats and plot}
}
\description{
functions that modified plot from raster package
}
\examples{
m <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/model.Rds"))


}
