% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_color.R
\name{points_color}
\alias{points_color}
\title{Plot points with color}
\usage{
points_color(
  x,
  y,
  z,
  col = hcl.colors(41, "Blue-Red"),
  zlim = range(z),
  pch = 19,
  cex = 1,
  outside = TRUE,
  verbose = F,
  ...
)
}
\arguments{
\item{x}{x-axis coordinate}

\item{y}{y-axis coordinate}

\item{z}{values}

\item{col}{color}

\item{zlim}{custom zlim}

\item{pch}{passed to points}

\item{cex}{passed to points}

\item{outside}{to include values outside zlim}

\item{verbose}{to display additional information}

\item{...}{other arguments passed to points and value2color}
}
\description{
Plot points with color
}
\examples{
library(raster)
## some maps to include
masp   <- raster::shapefile(paste0(system.file("extdata",package="hackWRF"),
                            "/RMSP.shp"))
br     <- raster::shapefile(paste0(system.file("extdata",package="hackWRF"),
                            "/BR_states.shp"))

## site-list w/ lat-lon
stations <- readRDS(paste0(system.file("extdata",package="hackWRF"),
                    "/stations.Rds"))

## open a evaluation table
sample <- read.stat(paste0(system.file("extdata", package = "hackWRF"),
                           "/sample.csv"),verbose=TRUE)
row.names(sample) <- c("Americana","Campinas","Congonhas")

## combination of lat-lon and evaluation table
coord  <- latlon(sample, coord = stations)

## plot
plot(NA, xlim = c(-47.75,-45.5), ylim = c(-24.7,-22.2),
     xlab = 'lon',ylab = 'lat')
points_color(x = coord$lon, y = coord$lat,z = coord$MB)
lines(masp, col = 'gray');lines(br)

}
