% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_emissions.R
\name{project_emission}
\alias{project_emission}
\title{project emissions using growth factors.}
\usage{
project_emission(r, shape, factor = rep(1, nrow(shape)), background = 1)
}
\arguments{
\item{r}{raster object}

\item{shape}{sf object from a shapefile containing 5 regions}

\item{factor}{numeric vector with 5 growth factors}

\item{background}{growth factors to be used outside the domain (default is 1)}
}
\description{
project emissions using growth factors.
}
\examples{
regions <- sf::read_sf(paste0(system.file("extdata",package="hackWRF"),"/BR_regions.shp"))
factors <- c(1.1,1.5,2,1.5,0.75)
# using XLAT only for test purposis
x  <- eixport::wrf_raster(paste0(system.file("extdata",package="hackWRF"),"/wrfinput_d01"),'XLAT')
project_x <- project_emission(x,regions,factors)
plot_raster(project_x,proj = TRUE)
}
