% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.stats.R
\name{read.stat}
\alias{read.stat}
\title{Function to read stats and evaluation}
\usage{
read.stat(file, csv = T, sep = ";", dec = ".", verbose = FALSE, ...)
}
\arguments{
\item{file}{model data.frame}

\item{csv}{logical, use write.csv instead of write.table#'}

\item{sep}{the field separator string, passed to  read.table function}

\item{dec}{he string to use for decimal points, passed to read.table function}

\item{verbose}{display additional information}

\item{...}{arguments passed to read.table functions}
}
\description{
Function to read stats and evaluation output
}
\examples{
sample <- read.stat(file    = paste0(system.file("extdata", package = "hackWRF"),"/sample.txt"),
                    csv     = FALSE,
                    verbose = TRUE)

sample <- read.stat(file    = paste0(system.file("extdata", package = "hackWRF"),"/sample.csv"),
                    csv     = TRUE,
                    verbose = TRUE)

}
