% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_satellite.R
\name{read_satellite}
\alias{read_satellite}
\alias{read_MODIS}
\alias{read_MODIS_HDF}
\alias{read_MERGE}
\alias{read_g4}
\alias{read_CERES}
\alias{read_OMI}
\alias{read_AIRS}
\title{Functions to read satellite products in NetCDF/HDF(4) format}
\usage{
read_MODIS(file, var, verbose = T)

read_MODIS_HDF(file, var, verbose = T)

read_MERGE(file, var = "prec", verbose = T)

read_g4(file, var = NA, verbose = T)

read_CERES(file, var, month = 1, year = 2012, verbose = T)

read_OMI(file, var = NA, k = 2.69e+16, verbose = T)

read_AIRS(file, var, k = 2.69e+16, verbose = T)
}
\arguments{
\item{file}{file name}

\item{var}{variable name, it can open a menu, in read_g4 it can read the first variable by defoult}

\item{verbose}{set TRUE to display additional information}

\item{month}{integer 1-12 (CERES only)}

\item{year}{integer, YYYY format (CERES only)}

\item{k}{conversion factor, default is 2.69E16 (OMI only), see notes}
}
\value{
a raster object
}
\description{
function to read MODIS NetCDF

function to read MODIS products in HDF(4) format

function to read satellite products in NetCDF format

function to read g4 selection in NetCDF format

function to read CERES products in NetCDF format

function to read OMI products in NetCDF format

function to read AIRS products in NetCDF format
}
\note{
Conversion from DU to molecules / cm 2
}
