% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_evaluate.R
\name{sat_evaluation}
\alias{sat_evaluation}
\title{Functions to model evaluation using satellite}
\usage{
sat_evaluation(mo, ob, n = 6, min = NA, max = NA, verbose = T)
}
\arguments{
\item{mo}{raster with model}

\item{ob}{raster with observations}

\item{n}{number of points from the boundary removed, default is 5}

\item{min}{minimum value cutoff}

\item{max}{maximum value cutoff}

\item{verbose}{set TRUE to display additional information}
}
\value{
a data.frame
}
\description{
functions to evaluate the spatial performance using satellite
}
