% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvertedSelectByDate.R
\name{SelectByInvertedDate}
\alias{SelectByInvertedDate}
\title{Utility function to make it easier to remove select periods from a data frame}
\usage{
SelectByInvertedDate(
  mydata,
  start = "1/9/2008",
  end = "30/9/2008",
  year,
  month,
  day,
  hour,
  incert_na = TRUE,
  verbose = T
)
}
\arguments{
\item{mydata}{A data frame containing a date field in hourly or high resolution format.}

\item{start}{A start date string in the form d/m/yyyy e.g. “1/2/1999” or in ‘R’ format i.e. “YYYY-mm-dd”, “1999-02-01”}

\item{end}{See start for format.}

\item{year}{not used at the moment}

\item{month}{not used at the moment}

\item{day}{not used at the moment}

\item{hour}{not used at the moment}

\item{incert_na}{incert a NA between between the two parts}

\item{verbose}{to display additional information}
}
\description{
Utility function to make it easier to remove select periods from a data frame
}
\examples{
model <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/model.Rds"))
plot(model$SAndre~model$date, ty = 'l', col = '#777777', lwd = 2)

model2 <- SelectByInvertedDate(model, start = '2012-01-03', end = '2012-01-05')
lines(model2$SAndre~model2$date, col = '#44AA44', lwd = 2)

}
