% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Calculate evaluation statistics from 2 vectors}
\usage{
stats(
  mo,
  ob,
  spinup = 0,
  wd = FALSE,
  scatter = F,
  add = F,
  cor = "#FF000088",
  lim = NA,
  cutoff = 0,
  nobs = 8,
  verbose = T,
  ...
)
}
\arguments{
\item{mo}{model}

\item{ob}{observed data}

\item{spinup}{number of model points excluded to the statistic}

\item{wd}{logical, true for wind direction, see notes}

\item{scatter}{set TRUE to plot a scatter plot}

\item{add}{set TRUE to add the points to a scatter plot}

\item{cor}{color of scatterplot dots}

\item{lim}{scatter plot limits}

\item{cutoff}{(optionally the maximum) valid value for observation}

\item{nobs}{minimum number of observations}

\item{verbose}{display additional information}

\item{...}{extra arguments passed to scatter plot}
}
\description{
Calculate statistical indexes
}
\note{
MB and ME are calculated using Mughal et al. (2017) for wind direction
}
\examples{
model <- 1:100
data  <- model + rnorm(100,0.2)
stats(mo = model, ob = data, scatter = TRUE)

}
