% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizer.R
\name{summarizer}
\alias{summarizer}
\title{make a summary using row.names}
\usage{
summarizer(files, rm.last = T, verbose = T)
}
\arguments{
\item{files}{list of files generated using write.stat}

\item{rm.last}{passed to read.stat}

\item{verbose}{display additional information}
}
\description{
Function that open and summarize many stats by row.name
}
