% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_emission_24h.R
\name{summary_emission_24h}
\alias{summary_emission_24h}
\title{Summary the emission from two emission files (00Z and 12z) from WRF-Chem model}
\usage{
summary_emission_24h(file_00, file_12, ...)
}
\arguments{
\item{file_00}{file patch and name for 00z emission file}

\item{file_12}{file patch and name for 12z emission file}

\item{...}{arguments passed to summary_emission}
}
\value{
a data.frame containing the pollutant name, minimum, average, maximum, units, molar weight and total (in kt/year)
}
\description{
Calculate the summary for each specie of gas and aerosol in two emission file
}
\examples{

\dontrun{
summary_emission_24h(file_00 = 'wrfchemi_00z_d01',
                     file_12 = 'wrfchemi_12z_d01')

}

}
\seealso{
summary_emission
}
