% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_color.R
\name{value_color}
\alias{value_color}
\title{Get color values from a vector inside a regular interval}
\usage{
value_color(
  var,
  col = hcl.colors(41, "Blue-Red"),
  interval = range(var),
  n_classes = length(col),
  verbose = T
)
}
\arguments{
\item{var}{numeric vector}

\item{col}{colors}

\item{interval}{range of values for the color interval (values outside the range are mapped in extremes)}

\item{n_classes}{number of classes, see details}

\item{verbose}{to display additional information}
}
\description{
Get color values from a vector inside a regular interval
}
\details{
the default number of classes (n_classes) is the number of colors (col) or length(var) - 1 if length(var) <= n_classes
}
\examples{
x <- sin(pi/8 * 1:15)
barplot(x, col = value_color(x))
box()

}
