% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.stats.R
\name{write.stat}
\alias{write.stat}
\title{Functions to write stats and evaluation}
\usage{
write.stat(stat, file, sep = ";", dec = ".", verbose = FALSE, ...)
}
\arguments{
\item{stat}{observed data.frame}

\item{file}{model data.frame}

\item{sep}{the field separator string, passed to write.table function}

\item{dec}{he string to use for decimal points, passed to write.table function}

\item{verbose}{display additional information}

\item{...}{arguments passed to write.table functions}
}
\description{
Functions to write stats and evaluation output. If a file name ending with .csv
is provided to the function will save using write.csv otherwise the function write.table.
}
\examples{

sample <- read.stat(paste0(system.file("extdata", package = "hackWRF"),"/sample.csv"),
                    verbose = TRUE)
dir.create(file.path(tempdir(), "stats"))

write.stat(file    = paste0(file.path(tempdir(), "stats"),'/sample.txt'),
           stat    = sample,
           verbose = TRUE)

write.stat(file    = paste0(file.path(tempdir(), "stats"),'/sample.csv'),
           stat    = sample,
           verbose = TRUE)

}
