% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Filter table A to perform a fair comparison with table B}
\usage{
comparison(tabA, tabB, verbose = T)
}
\arguments{
\item{tabA}{model}

\item{tabB}{observed data}

\item{clean}{remove rows with zero observations}

\item{summaryze}{add a last line with the the average values and format the table}
}
\description{
Filter table A to perform a fair comparison with table B using row.names
}
\examples{
model <- 1:10
data  <- model + rnorm(10,0.2)
A     <- stats(mo = model, ob = data)
for(i in 2:10){
   data  <- model + rnorm(10,0.2)
   A     <- rbind(A,stats(mo = model, ob = data))
}
row.names(A) <- c('a0','a1','a3','b0','b1','b2','b3','d0','d1','d2')

data   <- model + rnorm(10,0.2)
B      <- stats(mo = model, ob = data)
for(i in 2:5){
    data   <- model + rnorm(10,0.2)
    B      <- rbind(B,stats(mo = model, ob = data))

}
row.names(B) <- c('a0','a1','z0','z1','z2')

comparison(A,B)
}
