% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_range.R
\name{legend_range}
\alias{legend_range}
\title{Get the distance in kilometers between two points.}
\usage{
legend_range(
  x,
  text.width = 6,
  dig = c(2, 2, 2),
  xjust = 0.5,
  horiz = TRUE,
  y.intersp = 0.1,
  x.intersp = 0.1
)
}
\arguments{
\item{x}{raster or array}

\item{text.width}{Longitude in decimals}

\item{dig}{vector with number of digits for plot}

\item{xjust}{passed to legend}

\item{horiz}{passed to legend}

\item{y.intersp}{passed to legend}

\item{x.intersp}{passed to legend}
}
\description{
Get the distance in kilometers between two points.
}
\note{
for use with raster use before any change of projection

text.width can vary depending on map dimensions
}
\examples{
library(raster)
br <- shapefile(paste0(system.file("extdata",package="hackWRF"),"/BR_estates.shp"),verbose=FALSE)
plot(br)
box()
latitude()
longitude()
grid()
x <- eixport::wrf_get(paste0(system.file("extdata",package="hackWRF"),"/wrfinput_d01"),'XLAT')
legend_range(x)

}
