% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_emissions.R
\name{merge_emission}
\alias{merge_emission}
\title{Function to merge two different emissions}
\usage{
merge_emission(
  background,
  auxiliar,
  output = NA,
  name,
  coef_backgroud = 1,
  coef_auxiliar = 1,
  aux_n2 = NA,
  coe_n2 = 1,
  plot = T,
  verbose = T,
  ...
)
}
\arguments{
\item{background}{main input file (WRF-Chem emission file)}

\item{auxiliar}{auxliar input file (WRF-Chem emission file)}

\item{output}{output file (WRF-Chem emission file), or NA to return the emission array}

\item{name}{variable/pollutant name}

\item{coef_backgroud}{coefficient to multiply main input}

\item{coef_auxiliar}{coefficient to multiply auxliar input}

\item{aux_n2}{extra auxliar input file (will be merged with auxliar)}

\item{coe_n2}{coefficient for to multiply extra auxliar input}

\item{plot}{to plot from output file}

\item{verbose}{display additional information}

\item{...}{param for plot function}
}
\value{
if output is NA, returns an array or a matrix
}
\description{
Merge different emissions, the main emission or background emission
(usually from a global inventory) and an auxiliary emission. The resulting emissions
has the values of the auxiliar emissions for grid points with emissions and background
emission for grid points with zero emissions.
}
\examples{
\dontrun{
NEW_e <- merge_emission(background = 'base/wrfchemi_00z_d01',
                        auxiliar   = 'extra/wrfchemi_00z_d01',
                        name       = 'E_NO')
}
}
