% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\title{Read and write metadata on a NetCDF file}
\usage{
meta(
  filename = NA,
  variable = 0,
  attname = NA,
  action = "read",
  value = NA,
  verbose = F
)
}
\arguments{
\item{filename}{file name}

\item{variable}{variable name, 0 to global and "?" to ask all names}

\item{attname}{attribute names}

\item{action}{Read or write attribute (NA for get all attnames)}

\item{value}{value to write}

\item{verbose}{display additional information}
}
\description{
Read and write metadata information of a NetCDF files
}
\examples{
nc <- paste0(system.file("extdata",package="hackWRF"),"/small.nc")
meta(nc)
meta(nc,attname = "Title")
meta(nc,variable = "?")
meta(nc,variable = "so2")
meta(nc,variable = "so2",attname = "long_name")

}
