% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{latitude}
\alias{latitude}
\alias{longitude}
\title{PLot latitude and longitude}
\usage{
latitude(int = 10, side = 1, lmin = -180, lmax = 180, ...)

longitude(int = 10, side = 2, lmin = -90, lmax = 90, ...)
}
\arguments{
\item{int}{interval in degrees}

\item{side}{side to plot, see axis}

\item{lmin}{condition (TRUE or FALSE) for combine x and y or return only x}

\item{lmax}{message in case of bind is false (no binding)}

\item{...}{additional arguments passed to axis function}
}
\description{
function to plot cool latitude and longitude.
labels for latitude range from 180ºW to 0 to 180ºE and longitude from 90ºS to 0 to 90ºN.
}
\section{Functions}{
\itemize{
\item \code{latitude}: cool latitude

\item \code{longitude}: cool longitude
}}

\examples{
library(raster)
br <- shapefile(paste0(system.file("extdata",package="hackWRF"),"/BR_states.shp"),verbose=FALSE)
plot(br)
box()
latitude()
longitude()
grid()

}
