% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_color.R
\name{points_color}
\alias{points_color}
\title{Plot points with color}
\usage{
points_color(
  x,
  y,
  z,
  col = hcl.colors(41, "Blue-Red"),
  zlim = range(z),
  pch = 19,
  cex = 1,
  outside = TRUE,
  verbose = F,
  ...
)
}
\arguments{
\item{x}{x-axis coordinate}

\item{y}{y-axis coordinate}

\item{z}{values}

\item{col}{color}

\item{zlim}{custom zlim}

\item{pch}{passed to points}

\item{cex}{passed to points}

\item{outside}{to include values outside zlim}

\item{verbose}{to display additional information}

\item{...}{other arguments passed to points and value2color}
}
\description{
Plot points with color
}
\examples{
x <- sin(pi/8 * 1:15)
barplot(x, col = value2color(x))
box()

}
