% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Calculate evaluation statistics from 2 vectors}
\usage{
stats(
  mo,
  ob,
  spinup = 0,
  scatter = F,
  add = F,
  cor = "#FF000088",
  lim = NA,
  cutoff = 0,
  verbose = T,
  ...
)
}
\arguments{
\item{mo}{model}

\item{ob}{observed data}

\item{spinup}{numer of model points excluded to the statistic}

\item{scatter}{set TRUE to plot a scatter plot}

\item{add}{set TRUE to add the points to a scatter plot}

\item{cor}{color of scatterplot dots}

\item{lim}{scatter plot limits}

\item{...}{extra arguments passed to scatter plot}
}
\description{
Calculate statistical indexes
}
\examples{
model <- 1:100
data  <- model + rnorm(100,0.2)
stats(mo = model, ob = data, scatter = TRUE)

}
