% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_emission.R
\name{summary_emission}
\alias{summary_emission}
\title{Summary the emission from a emission file from WRF-Chem model}
\usage{
summary_emission(
  file = choose.files(),
  chem_opt = "cb05",
  remove_zeros = TRUE,
  remove_dim = TRUE,
  pm = FALSE,
  voc = FALSE,
  mask = NULL,
  verbose = F
)
}
\arguments{
\item{file}{file patch and name}

\item{chem_opt}{chemical mechanism option, default is CB05}

\item{remove_zeros}{remove empty values (all zeros) from table, default is TRUE}

\item{remove_dim}{remove dimension variables XLAT, XLON and Times from table,default is TRUE}

\item{pm}{calculate total PM, default is FALSE}

\item{voc}{calculate total voc, nox and nox/voc ratio and nox/voc ratio in propylene-eq}

\item{mask}{sf object MULTIPOLYGON from a shapefile, see notes and example}

\item{verbose}{display additional information}
}
\value{
a data.frame containing the pollutant name, minimum, average, maximum, units, molar weight and total (in kt/year)
}
\description{
Calculate the summary for each specie of gas and aerosol in a given emission file
}
\note{
Molar mass for PAR is estimated considering weighted average of: Propane 11\%, Butanes 18\%, Pentanes 9\%, Hexanes and higher alkanes 39\%, Acetylene 5\%, Benzene 11\%, Esters 3\%, Ethers 4\%) divided by the number of carbons (1.5 for propane, 4 for Butanes, 5 for pentanes, 5.86 for Hexanes and 1 for the others).

If the mask argument is provided, only the inner region of the shapefile is used to determine the total emissions.
}
\examples{
# sample shapefile for Brazil and Brazilian regions
BR         <- sf::read_sf(paste0(system.file("extdata",package="hackWRF"),"/BR.shp"))
BR_regions <- sf::read_sf(paste0(system.file("extdata",package="hackWRF"),"/BR_regions.shp"))

\dontrun{
file <- choose.files()

# fast check for the entire file
summary_emission(file = file)

# check using the BR
summary_emission(file = file,mask = BR)

# check using the midwest region from BR_regions
midwest <- BR_regions[2,]
summary_emission(file = file,mask = midwest)

# opening map for Metropolitan area of Sao Paulo
masp <- sf::read_sf(paste0(system.file("extdata",package="hackWRF"),"/RMSP.shp"))
# comnining the citis into a mask
masp <- sf::st_combine(sf::st_union(masp))
summary_emission(file = file,mask = masp)
}

}
